RB1-2S5BAK1A: fix shield hole diameter
At first, I was drawing footprint for other UDE connector, which was
generally very similar, but its shielding hole was advised as 1.5 mm. In
case of this connector it should be 1.6 mm. This overlook most likely
causes issues during mounting. It fits very tightly, which makes factory
usage rather impossible. After this fix, it should be way better.